feat(models): add Sequelize configuration and initialize models
