--- src/models/index.ts
@@ -1,60 +1,0 @@
-"use strict";
-
-import * as fs from "fs";
-import * as path from "path";
-import { Sequelize, DataTypes } from "sequelize";
-import * as process from "process";
-
-// Interface para o objeto de banco de dados
-interface DB {
-  sequelize: Sequelize;
-  Sequelize: typeof Sequelize;
-  [modelName: string]: any;
-}
-
-const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || "development";
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const config = require(__dirname + "/../config/config.json")[env];
-const db: DB = {} as DB;
-
-let sequelize: Sequelize;
-if (config.use_env_variable) {
-  sequelize = new Sequelize(
-    process.env[config.use_env_variable] as string,
-    config
-  );
-} else {
-  sequelize = new Sequelize(
-    config.database,
-    config.username,
-    config.password,
-    config
-  );
-}
-
-fs.readdirSync(__dirname)
-  .filter((file: string) => {
-    return (
-      file.indexOf(".") !== 0 &&
-      file !== basename &&
-      (file.slice(-3) === ".js" || file.slice(-3) === ".ts") &&
-      file.indexOf(".test.") === -1
-    );
-  })
-  .forEach((file: string) => {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const model = require(path.join(__dirname, file))(sequelize, DataTypes);
-    db[model.name] = model;
-  });
-
-Object.keys(db).forEach((modelName) => {
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
-});
-
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
-
-export default db;
